bank_api/pyproject.toml
[tool.poetry]
name = "bank"
version = "0.1.0"
description = ""
authors = ["Rojen Maharjan <44132059+rojen11@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
sqlalchemy = "^2.0.19"
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.1.1"}
psycopg2-binary = "^2.9.7"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
